+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Added the attribute `attr(gom, "npy")` (with value 3) to the `gom` dataset.  This is for compatability with the **threshr** package.
 
+* Give an explicit error message if `plot.evpost` is called with the logically incompatible arguments `add_pu = TRUE` and `pu_only = TRUE`.
+
 # revdbayes 1.2.1
 
 ## Bug fixes and minor improvements